Move legacy sidebar JS into a separate module
