+++ client/src/entrypoints/admin/wagtailadmin.js
@@ -3,10 +3,7 @@ import ReactDOM from 'react-dom';
 import {
   Icon,
   Portal,
-  initExplorer,
   initFocusOutline,
-  initSubmenus,
-  initSkipLink,
   initIE11Warning,
   initUpgradeNotification,
 } from '../..';
@@ -29,16 +26,7 @@ window.wagtail.components = {
  * Add in here code to run once the page is loaded.
  */
 document.addEventListener('DOMContentLoaded', () => {
-  const explorerNode = document.querySelector('[data-explorer-menu]');
-  const toggleNode = document.querySelector('[data-explorer-start-page]');
-
-  if (explorerNode && toggleNode) {
-    initExplorer(explorerNode, toggleNode);
-  }
-
   initFocusOutline();
-  initSubmenus();
   initIE11Warning();
   initUpgradeNotification();
-  initSkipLink();
 });
